gwas workflow debug; incoporated input() into GProcessor

+++ heig/wgs/gwas.py
@@ -2,7 +2,13 @@ import os
 import shutil
 import hail as hl
 import heig.input.dataset as ds
-from heig.wgs.utils import GProcessor, keep_ldrs
+from heig.wgs.utils import GProcessor
+
+
+"""
+TODO: consider providing more preprocessing options? such as --chr
+
+"""
 
 
 def config(args):
@@ -38,20 +44,15 @@ def check_input(args, log):
     if args.geno_mt is not None and not os.path.exists(args.geno_mt):
         raise FileNotFoundError(f"{args.geno_mt} does not exist")
 
-    # optional arguments
-    # if args.mem is None:
-    #     args.mem = 8
-    # if args.threads is None:
-    #     args.threads = 4
     if args.n_ldrs is not None and args.n_ldrs <= 0:
         raise ValueError('--n-ldrs should be greater than 0')
     
     if args.maf_min is not None:
-        if args.maf_min >= 0.5 or args.maf_min <= 0:
+        if args.maf_min > 0.5 or args.maf_min < 0:
             raise ValueError('--maf-min must be greater than 0 and less than 0.5')
     else:
-        args.maf_min = 0.01
-        log.info(f"Set --maf-min as default 0.01")
+        args.maf_min = 0
+        # log.info(f"Set --maf-min as default 0.01")
     
     if args.variant_type is None:
         args.variant_type = 'variant'
@@ -62,26 +63,29 @@ def check_input(args, log):
         
     if args.maf_max is None:
         args.maf_max = 0.5
-    elif args.maf_max >= 0.5 or args.maf_max <= 0 or args.maf_max <= args.maf_min:
+    elif args.maf_max > 0.5 or args.maf_max <= 0 or args.maf_max <= args.maf_min:
         raise ValueError(('--maf-max must be greater than 0, less than 0.5, '
                           'and greater than --maf-min'))
     
     # process arguments
-    try:
-        start, end = args.range.split(',')
-        start_chr, start_pos = [int(x) for x in start.split(':')]
-        end_chr, end_pos = [int(x) for x in end.split(':')]
-    except:
-        raise ValueError(
-            '--range should be in this format: <CHR>:<POS1>,<CHR>:<POS2>')
-    if start_chr != end_chr:
-        raise ValueError((f'starting with chromosome {start_chr} '
-                            f'while ending with chromosome {end_chr} '
-                            'is not allowed'))
-    if start_pos > end_pos:
-        raise ValueError((f'starting with {start_pos} '
-                            f'while ending with position is {end_pos} '
-                            'is not allowed'))
+    if args.range is not None:
+        try:
+            start, end = args.range.split(',')
+            start_chr, start_pos = [int(x) for x in start.split(':')]
+            end_chr, end_pos = [int(x) for x in end.split(':')]
+        except:
+            raise ValueError(
+                '--range should be in this format: <CHR>:<POS1>,<CHR>:<POS2>')
+        if start_chr != end_chr:
+            raise ValueError((f'starting with chromosome {start_chr} '
+                                f'while ending with chromosome {end_chr} '
+                                'is not allowed'))
+        if start_pos > end_pos:
+            raise ValueError((f'starting with {start_pos} '
+                                f'while ending with position is {end_pos} '
+                                'is not allowed'))
+    else:
+        start_chr, start_pos, end_pos = None, None, None
 
     temp_path = 'temp'
     i = 0
@@ -99,15 +103,25 @@ def check_input(args, log):
 
 
 def do_gwas(snps_mt, n_ldrs, n_covar, log):
-    pheno_list = [snps_mt.ldrs[i] for i in range(n_ldrs - 1)]
-    covar_list = [snps_mt.covar[i] for i in range(1, n_covar - 1)]
-    covar_list.append(1)
+    pheno_list = [snps_mt.ldrs[i] for i in range(n_ldrs)]
+    covar_list = [snps_mt.covar[i] for i in range(n_covar)]
 
     log.info(f'Doing GWAS for {n_ldrs} LDRs ...')
-    gwas = hl.linear_regression_rows(y=pheno_list,
-                                     x=snps_mt.GT.n_alt_alleles(),
-                                     covariates=covar_list,
-                                     pass_through=['rsid'])
+    gwas = hl.linear_regression_rows(
+        y=pheno_list, x=snps_mt.GT.n_alt_alleles(),
+        covariates=covar_list, pass_through=[snps_mt.rsid, snps_mt.info.n_called]
+    )
+    # gwas = gwas.annotate(n_called=snps_mt.index_rows(gwas.key).info.n_called)
+    gwas = gwas.annotate(chr=gwas.locus.contig,
+                         pos=gwas.locus.position,
+                         ref_allele=gwas.alleles[0],
+                         alt_allele=gwas.alleles[1])
+    gwas = gwas.key_by() 
+    gwas = gwas.drop(*['locus', 'alleles', 'y_transpose_x', 'sum_x', 'n'])
+    gwas = gwas.select('chr', 'pos', 'rsid', 'ref_allele', 'alt_allele', 
+                       'n_called', 'beta', 'standard_error', 't_stat',
+                       'p_value')
+
     return gwas
 
 
@@ -117,11 +131,15 @@ def run(args, log):
     hl.init(quiet=True)
     hl.default_reference = geno_ref
 
+    log.info(f'Read LDRs from {args.ldrs}')
     ldrs = ds.Dataset(args.ldrs)
-    log.info(f'{ldrs.data.shape[1]} LDRs read from {args.ldrs}')
+    log.info(f'{ldrs.data.shape[1]} LDRs and {ldrs.data.shape[0]} subjects.')
     if args.n_ldrs is not None:
-        ldrs.data, _ = keep_ldrs(args.n_ldrs, ldrs.data)
-        log.info(f'Keep the top {args.n_ldrs} LDRs.')        
+        ldrs.data = ldrs.data.iloc[:, :args.n_ldrs]
+        if ldrs.data.shape[1] > args.n_ldrs:
+            log.info(f'WARNING: --n-ldrs greater than #LDRs, use all LDRs.')
+        else:
+            log.info(f'Keep the top {args.n_ldrs} LDRs.')        
 
     log.info(f'Read covariates from {args.covar}')
     covar = ds.Covar(args.covar, args.cat_covar_list)
@@ -144,25 +162,25 @@ def run(args, log):
     # snps_mt
     if args.bfile is not None:
         log.info(f'Read bfile from {args.bfile}')
-        snps_mt = hl.import_plink(bed=args.bfile + '.bed',
-                                  bim=args.bfile + '.bim',
-                                  fam=args.bfile + '.fam',
-                                  reference_genome=geno_ref)
+        gprocessor = GProcessor.import_plink(args.bfile, geno_ref,
+                                             variant_type=args.variant_type, 
+                                             maf_min=args.maf_min, maf_max=args.maf_max)
     elif args.geno_mt is not None:
         log.info(f'Read genotype data from {args.geno_mt}')
-        snps_mt = hl.read_matrix_table(f"{args.geno_mt}")
-
-    gprocessor = GProcessor(snps_mt, geno_ref=geno_ref, 
-                            variant_type=args.variant_type, 
-                             maf_min=args.maf_min, maf_max=args.maf_max)
+        gprocessor = GProcessor.read_matrix_table(args.geno_mt, geno_ref,
+                                                  variant_type=args.variant_type, 
+                                                  maf_min=args.maf_min, maf_max=args.maf_max)
+   
     log.info(f"Processing genetic data ...")
     gprocessor.extract_snps(keep_snps)
     gprocessor.extract_idvs(common_ids)
     gprocessor.do_processing(mode='gwas')
-    gprocessor.extract_gene(chr=chr, start=start, end=end)
+    if chr is not None and start is not None:
+        gprocessor.extract_gene(chr=chr, start=start, end=end)
     
-    log.info(f'Save preprocessed genotype data to {temp_path}')
-    gprocessor.save_interim_data(temp_path)
+    if not args.not_save_genotype_data:
+        log.info(f'Save preprocessed genotype data to {temp_path}')
+        gprocessor.save_interim_data(temp_path)
     gprocessor.check_valid()
     
     try:
@@ -173,29 +191,30 @@ def run(args, log):
         covar.keep(snps_mt_ids)
         covar.cat_covar_intercept()
         log.info(f'{len(common_ids)} common subjects in the data.')
-        log.info(f"{covar.shape[1]} fixed effects in the covariates after removing redundant effects.\n")
+        log.info(f"{covar.data.shape[1]} fixed effects in the covariates (including the intercept).")
 
-        covar.data.to_csv(f'{temp_path}_covar.tb', sep='\t')
-        covar_table = hl.import_table(f'{temp_path}_covar.tb', key='IID', impute=True, types={'IID': hl.tstr})
-        ldrs.data.to_csv(f'{temp_path}_ldrs.tb', sep='\t')
-        ldrs_table = hl.import_table(f'{temp_path}_ldrs.tb', key='IID', impute=True, types={'IID': hl.tstr})
+        covar.data.to_csv(f'{temp_path}_covar.txt', sep='\t')
+        covar_table = hl.import_table(f'{temp_path}_covar.txt', key='IID', impute=True, types={'IID': hl.tstr})
+        ldrs.data.to_csv(f'{temp_path}_ldrs.txt', sep='\t')
+        ldrs_table = hl.import_table(f'{temp_path}_ldrs.txt', key='IID', impute=True, types={'IID': hl.tstr})
 
         # annotate ldrs and covar to snps_mt
         gprocessor.annotate_cols(ldrs_table, 'ldrs')
         gprocessor.annotate_cols(covar_table, 'covar')
-        # snps = snps.annotate_cols(ldrs=ldrs_table[snps.s])
-        # snps = snps.annotate_cols(covar=covar_table[snps.s])
 
         # gwas
         n_ldrs = ldrs.data.shape[1]
         n_covar = covar.data.shape[1]
         gwas = do_gwas(gprocessor.snps_mt, n_ldrs, n_covar, log)
+        
+        ## 1:03 continue checking this
 
         # save gwas results
         log.info(f"Save GWAS results to {args.out}.txt.bgz")
         gwas.export(f"{args.out}.txt.bgz")
     finally:
-        shutil.rmtree(temp_path)
-        shutil.rmtree(f'{temp_path}_covar.tb')
-        shutil.rmtree(f'{temp_path}_ldrs.tb')
-        log.info(f'Removed preprocessed genotype data at {temp_path}')
+        if os.path.exists(temp_path):
+            shutil.rmtree(temp_path)
+        # shutil.rmtree(f'{temp_path}_covar.txt')
+        # shutil.rmtree(f'{temp_path}_ldrs.txt')
+            log.info(f'Removed preprocessed genotype data at {temp_path}')
